requirements.txt
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions>=0.59b0

